Made the icons in the main activity react to onPressed.
In order to do this, I had to move the onClickListener from the layout
to the button. This is because the button doesn't react to an onClick in
the layout, and if the button is set to clickable, it will steal the
clicks from the layout.

The downside is that the clicking area became smaller, but I feel that
users are clicking the button anyways.