Remove useless intent filters for in-application activities
As stated in the Android documentation, "an explicit intent is always
delivered to its target, no matter what it contains; the filter is not
consulted".

There are no advantage in letting those filters in, as they cannot
match since they lack a category filter, and every implicit filter
have at least one category.