Cache ETag and Last-Modified HTTP headers for images
Instead of comparing the "Last-Modified" dates and adding a round-trip to
the server, this now caches the "ETag" and "Last-Modified" headers if
they are present in the server HTTP response.

Those values are given back (using respectively "If-None-Match" and
"If-Modified-Since" request headers) to the server when requesting an
update. Most servers will answer with a "304" (use local copy) if the
requested resource has not been updated. In the unlikely case that the
server does not honor those headers, the image will be refreshed in any
case, so we will not serve old images to the user.

The rationale for storing the "Last-Modified" return value is to conform
to RFC 2616 advice to treat it as opaque because the server might do an
exact string comparison instead of an ordered date comparison.