Refactoring: remove useless null checks
Boundary data ultimately coming to loadInViewport is checked for null
(those are Long fields) while they come from long values, that cannot be
null.

It makes little sense to return null to represent an empty set when we can
return the empty set itself.

Also, the cursor cannot be null, and the method can be made private.