fix #1347: Unit tests fail on Intel image with VM acceleration
- skipping localization of enums when cgeoapplication.getInstance() is
  null
- don't use cgeo logging but plain android logging in test cases
  (cgeo logging needs Settings which needs cgeoapplication)

Enums access cgeoapplication object in set10n() but in AndroidTestCases
the application object is not guaranteed to be initialized when the
test runs. This leads to NPEs when using cgeoapplication.getInstance().

On ARM emulator or during debugging, cgeoapplication is created
early enough and tests are green (by chance) but this is not the case on
Intel emulator with VM acceleration. Here, tests like CacheTest fail
and succeed randomly.