Simplify previous fix due to root cause fixing
The fix for #1680 introduced in ad93745d42f9aaf5bc4925075987ef64c13b2f71
had to account for the possibility of angles being outside the [0, 360[
range.

Commit 0034e358668d5911639628468be7409b2696d6c5 fixed #1685 and now
ensures that the returned orientation is within the right domain range
even after the phone rotation has been added to the sensor angle.

It is now possible to simplify the computation because we now that the
difference of two angle values will always be in the ]-360; 720[ range,
and can this fixed by adding 360 and taking the module with 360 to enter
back the [0, 360[ range.