Update main/res/values-it/strings.xml
As an Italian user, I checked out the Italian rules for the "gender" of the cache. 
In Italian the rule is that when a word doesn't have a translation, it should be used as a "male" word. In addition the italian synonymous of the word cache (or geocacher) are "tesoro", "nascondiglio" or even "contenitore" that are all "male".
Official application on Android (and iPhone and Windows Mobile) are using that word as a "male".
Plus, the italian dubbers of two episodes of Law & Order where it is mentioned the goeocaching regers to "geocache" as a "male". And a german translator that is working on translating a detective book that involves geocaching has used "cache" as a "male".