Fixes #3000: Caches missing on map (basic members)
Keeps a list of caches with no coordinates loaded (if user is not premium
and has not enabled Recaptcha) and removes them to force a reload.

Takes a moment after all the other caches load, but all those caches which were
previously loaded on the Nearby list will load on the Live Map for basic
members now.