fix #3480: quasi-systematic concurrent logins
The login procedure was initiated in both `onCreate()` and `onResume()`.
It has been limited to `onResume()` and refactored as to not manipulate
a global variable.

Also, the previous code prevented cgeo from retrying a failed login to
one of the cache providers if another had succeeded, as every provider
cleared `setLogin()` in case of success. This has been fixed as well.