Update for RxJava 0.17.0-RC1
This is the release candidate version, which is being currently tested
at Netflix. In a few days, I'll update with the final version.

This release cleans up some bits of RxJava by integrating the
subscription inside a subscriber (which is an observer holding a
subscription) and fixes a leak with recurring tasks.