Favor the orientation sensor over the rotation one
On some devices, it looks like the orientation sensor may be bogus. This
confirms what the author of GPS Status wrote in private mail.

This commits favors the rotation sensor over the orientation one, unless
low-power mode is used, in which case the orientation sensor (which may
benefit from low-power mode if it has a gyroscope-less mode) will be
preferred.

Part of work on #4599.