fix #4596: UI delay when building the list of caches to refresh
The list of caches to refresh was partitioned into caches already having
static maps and others, the later being prioritized in the refresh. This
computation was done on the UI thread, which led to a pause for large
lists.

The list building in now done in the background, with emission of caches
to refresh as soon as possible to reduce the user wait.