Fix #4858: TravelBug doesn't have good icon in CacheDetailsActivity inventory
The list in CacheDetailsActivity is filled from the cache page. In this
page, there is only de guid information, no Trackable serial number.

When trackable are parsed in GCParser, they are stored in the Datastore.
In this parsing, Trackable brand is set.

Save Trackable brand in datastore.
So Travelbug will have the good logo instead of unknown.