Fix #4806: GK requests to GC
The TrackableActivity try every connector and return the first one matching
the pattern. TravelBugConnector also match GK numbers as Tracking Code search.

This patch limit 'bad' requests made to GC services. It ONLY works if we open
the TrackableActivity from CacheDetailActivity as we know the brand of the
trackable we would open. This is not the case when opening trackableActivity
from the 'search bar', which is good. Issue #4822 still happens in this case.