Fix #5193: Icon size in lists different with/without overlay
I think the original icon size was enlarged in map the day overlays were
added around a cache. But the original marker icon size wasn't updated.
In lists, cache icon keeped their original size until #1210 alignment of
icons between live-map and lists. Moreover, resolution detection was wrong
since a long time. LDPI and MDPI shared a resolution of 0. And all other
resolutions used the n-1 resolution. This is a bug fixed by #aa58877. This
has reveled that the visual appearence "was good by mistake". In order
to keep the same visual appearence, this add a "fake" layer which force
markers in map to keep their "usual" size. In lists, adding the overlays
forced to enlage the cache type icon the same as in the map. This also set
them to a coherent size, even when no overlays are present. Cache type icons
are now always larger than before #1210.