fix #6284 log date format in system information
Changed `Log date format` in SystemInformation to `System date format` to make it clear what it is.
Used a cast to SimpleDateFormat to retrieve the data format pattern used.
Backed by a UnitTest.

I also resolved a warning in the `Formatter` class about a possible memory leak because the Context was stored in a static field.