Improved string
The "Duplicate" function was previously only called by a waypoint options menu, but now it can also be called by clicking on the coords of an existing waypoint. Therefore it is better to make clear the "duplicate" means duplicating the whole waypoint.

This change should reset and retrigger translations on Crowdin automatically on next string upload to Crowdin (@rsudev Feel free).