update to android-gradle 2.3
This is the most recent version of the gradle build support for Android.
It also requires updating SDK build to 25.0 or better. So it seems
reasonable to take the latest version of SDK build.

If you see local warnings, please also upgrade Android Studio itself to
2.3.

This change probably requires Jenkins nodes to be upgraded.