rename Eclipse cgeo project to main
Up to now the application project was called cgeo in Eclipse (since the
beginning of time). This change renames the project to main, and adapts
all the Eclipse related settings. This makes the logical project
structure more similar in both Eclipse and IDEA. It should also simplify
the use of gradle tooling in Eclipse, since that seems to be confused by
using a different logical name than the real file name.

There are absolutely no changes in the real directory structure for user
of other IDEs or build systems.

If you see errors in Eclipse after pulling this change, remove the cgeo
project from the workspace (without deleting it from disk) and restart.
The main project should be imported automatically now.