fix #6715 merge WP userNotes with Calculator notes
Current data in the calculator notes is simply ignored. Possible data loss for nightly build users! But the calculator variables are still available, so the loss might not be too hard.

It also removes the calculate button from CoordinatesInputDialog when called from NavigateAnyPointActivity, SearchActivity and LogTrackableActivity. That caused a crash.