fix #6826 remove superfluous waypoint note
For user defined waypoints, the note is merged with the userNote and the original note be emptied.

Empty note fields get hidden, even for server defined waypoints.

For server defined notes gets preserved. They are not editable but we stripped html. Now html can be preserved.

Html in userNotes is not supported and "<" characters escaped.

This might solve #6791 as well.