fix #6861 support null coords in waypoints
The EditWaypointActivity is not anymore prefilled with the users location. Placeholders are visible on the latitude and longitude buttons.
If the user just wants to have the current location he has to open the CoordsInputDialog once and confirm.
The projection is only enabled if coordinates are set on the waypoint. Otherwise the projection fields are disabled. We can't do a projection from null coordinates.
I've also added a "Clear Coordinates" button on the CoordsInputDialog for convenience.
Invalid coordinates from the Calculator are now forwarded as null to the waypoint. No longer the users location is used in the meantime. This solves #6697 for me. I don't know if someone needs more visual clue.

Unfortunately the "simple" usecases like recording my current position as waypoint or a simple projection from my location needs two more clicks now.

What gets more apparent now is the fact that the CoordsInputDialog is in between the EditWaypointActivity and the Calculator. You can have null coordinates now in the EditWaypointActivity because of an invalid calculation, but when you navigate through the CoordsInputDialog you see your current location as coordinates but don't use it.

Maybe a future refactoring for e.g. #6833 can simplify the dialog(s).