fix #6898 coords input zero padding
I tried to fix this, so that no zero padding happens on format switch when there is no input at all in the fields. But padding on input done (check mark). The only fields I left unpadded are the degree fields. I believe these are the only mandatory fields of a coordinate which shouldn't be set to all zeros when left empty.