fix #5758 and fix #5757 drop from temp list
Temporary lists didn't offer the "drop" action, even if the list contains stored caches.
This change makes the "drop" action available for temporary lists, like a list created from the map (view as list), if some of the caches are stored. It drops the caches from the device (remove from any list).