fix #6973 log "Report Problem" a.k.a NM, NA
Implements the "Report Problem" for the new GC logging page and also for OC (#5539).

It partly solves #4899 for those cases where a cacher wants to save e.g. a found and a needs maintenance offline or a DNF and a needs archived.

It also solves #6812 because I had to change the logic to save the log when the report problem changed.