Fix cgeo#4697: TB selections and GC vote lost when returning to logging activity
After returning from TB detail view - do not reload the log data when already TB data exist.
This is a little bit dirty, but I think we call this only in this situation.

Additional - deactivate the "AndroidRxUtils.bindActivity().subscribe()" in onCreate().
In my tests this code was not necessary to work correct.