use task avoidance gradle API
Gradle is changing from immediate to lazy task creation and
configuration for better performance. This requires new API to be used
in build scripts.

This change converts some of our tasks. Unfortunately, most of the
Android plugins tasks are still created eagerly, so we only avoid a
small number of tasks to be created.