Create a "Getting Started" page to display on first installation, deactivate geocaching.com on start.
fix #7461

I have created a "Getting Started" page that is displayed to the user the first time they start after installation. An initial text is defined. Of course, this is a first idea now. I think that there may be additional links in other languages, e.g. in German language to https://www.cachewiki.de/wiki/Hauptseite.

The page is presented before the permissions questions (on map access).