Create a "Getting Started" page to display on first installation
fix#4263 (1st impression in application is bad)

As before discussed in #7461 now separated the idea for an initial welcome screen.

I have created a "Getting Started" page that is displayed to the user the first time c:geo start after installation.
It is located in the "about c:geo" section, so the user can read it also later.
The initial text is split in 3 section - general, permission and platform information.
This could be extended if necessary, e.g. in German language to https://www.cachewiki.de/wiki/Hauptseite.

The page is presented before the permissions questions (on map access).