Update implemented CacheAttributes according to current OC acodes
fix #7671 (Missing attribute symbol for OR00C1 (from oc.ro))

readme.txt
- Update for okapi attributes according to the current code tree and
wiki

attribute-definitions.xml
- replace by the current used okapi version

attrlist.txt
- update is generated from the current attribute-definitions.xml
- Hint: New in the current OKAPI attribute-definitions.xml are acode A74
- A77
  * Defined in the iconlist.txt/CacheAttribute.java are the values for
  A74 and A75
    * A76 and A77 (OCUK attributes) need to be defined in a PR for new
    values (Issue #7705)

iconlist.txt
- update the iconlist according to the attribute definitions

CacheAttribute.java
- one line for every attribute
- configure OC acode numbers (A74, A75) to the alredy defined
    attributes

The acodes greater A77 (see in the wiki) are not in the current
OKAPI attribute-definitions.xml - the are work in progress in
opencaching/okapi#590 and opencaching/okapi#591. We should integrate
this in c:geo after the PR is merged. (Issue #7705 or later)

The code refresh for the generating scripts (*.sh, *.jar) will be
done in the planed PR to #7705