Fix #9069 - Hint is not displayed on compass view, when launched directly from map
When gathering from another cache, if this cache wasn't detailed, and the other is,
the hint, etc. would never be copied. This fixes the case where the cache is stored
in the database.

Also, if there isn't a hint (or it isn't stored), don't show the show-hint icon on
the compass UI.