add missing OC attributes
fixes #7705

In the part for `attributes_okapi` I decided to remove the old eclipse configuration
and switch to manual compilation for the attribute_definition.xml analyze tool.

This tool is updated for all OC location and results in more information.

The readme.txt covers the changes.

The attribute_definition.xml was updated from the current okapi version.

In the part `attributes` the icon workflow was refreshed.
This is covert in the readme.txt there.

The shell scripts got some code cleanup.
Only the cache icon overview script changes to present the acode instead of the OCDE code.

The OCDE icons overview file was refreshed and renamed to make clear the OCDE reference.
It got two additional attribute lines.

For the 8 new OC attributes I created the attribute icons based on PD/CC licenced vector images.