use HtmlCompat.fromHtml instead of deprecated Html.fromHtml
Since API24 the plain Html.fromHtml(string) is deprecated in favor of Html.fromHtml(string, int).
There is a wrapper class HtmlCompat where the method is switch according to the Build version of the device is done.
The implementation uses now only the int FROM_HTML_MODE_LEGACY (value 0).

Maybe we can optimize later the code using some of the new parameters to solve issue #7306.