replace icon-creation by ribbonizer by dedicated (ribbonized) images
The icon which is to change is primary defined in the AndroidManifest.xml.
But also it is used in the about screen (and now in the configuration wizard).

To achive a dedicated icon for different build types we have two possibilities.
At first we can let substitute this (and this is also possible for other entries in AndroidManifest) by declaring a manifestPlaceholders definition in the gradle buildTypes definition, e.g. `manifestPlaceholders = [appIcon: "@mipmap/ic_launcher"]` and using it         `android:icon="${appIcon}"`.
This would be a little code change. But in this way we can only manipulate the AndroidManifest.xml.
The other way is the possibility to define an additional `res` directory for the build type(s). There we need only the changed files. The big advantage is that we can refer to the changed ressource in the whole project (what we need for the about screen).
The main code change is that we have to move the main ressource directory to his "normal" place (from Android point of view).

I decided to go the second way in the PR.

Additionals:
- The management for keys.xml was adopted.
- The res folder mipmap-anydpi-v26 has xml files without ribbon, therefore saved to rawimages.
  Maybe we can create adaptove icons later with ribbon (or another colored? difference).

For the ribbon I used different colors and strings to easier differentiate the build types. Used rc/beta=red, nightly=black, debug=blue. The legacy build type has no dedicated icon.
This is open to discussion.