AGP 7.x
Move AGP from 4.2.2 to 7.1.3, partially using the assistent.
Renaming `lintOptions` to `lint` as required from AGP.

The lint tool results in
~~~
res\values-fr\strings.xml:347: Error: For locale "fr" (French) the following quantities should also be defined: many [MissingQuantity]
    <plurals name="caches_more_caches_next_x">
~~~
As `plurals` is not supported on crowdin we have to disable it.