gradle 7.6.3
Update gradle to the latest in the 7.6 serie: 7.6.3.
Activating the (incubating) configuration cache. Externalize File declarations from tasks in the gradle file to work with configuration caching.

For release notes see
 - https://docs.gradle.org/6.3/release-notes.html
 - https://docs.gradle.org/7.6.3/release-notes.html

The new parameter "networkTimeout" in the gradle-wrapper.properties allows now to change the value to other than 10 seconds (the default).